% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalExtractUtils.R
\name{acosd}
\alias{acosd}
\title{Arccosine in Degrees}
\usage{
acosd(x)
}
\arguments{
\item{x}{A numeric value or vector. The input should be within the range [-1, 1], as values outside this range will produce NaN values.}
}
\value{
A numeric value or vector representing the arccosine of the input, expressed in degrees.
}
\description{
The `acosd` function computes the inverse cosine (arccosine) of a numeric value and converts the result from radians to degrees.
}
