% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointStrucEstUtils.R
\name{ccpOutAnalysisMJ}
\alias{ccpOutAnalysisMJ}
\title{Analyze Optimization Output by Calculating Projection Angles}
\usage{
ccpOutAnalysisMJ(cache_v, VBars)
}
\arguments{
\item{cache_v}{A list of optimization vectors obtained from iterations of the optimization algorithm.}

\item{VBars}{A list of adjusted signal row spaces (matrices).}
}
\value{
A list of length equal to the number of signal row spaces. Each element is a numeric vector of
angles for each iteration of the optimization process.
}
\description{
The function `ccpOutAnalysisMJ` computes the projected angles between each cached optimization vector
and each adjusted signal row space. The projection angle serves as a metric for evaluating alignment
between the vectors obtained during optimization and the reference row spaces.
}
