% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointStrucEstUtils.R
\name{ccpOutVisualMJ}
\alias{ccpOutVisualMJ}
\title{Visualize the Optimization Progress of Projected Angles}
\usage{
ccpOutVisualMJ(
  angleHats,
  phiBars,
  dataname,
  iprint = FALSE,
  figdir = NULL,
  figname = NULL
)
}
\arguments{
\item{angleHats}{A list of numeric vectors representing the projected angles of each data block at different iterations.}

\item{phiBars}{A numeric vector representing the perturbation angles for each data matrix.}

\item{dataname}{A character vector containing the names of the data blocks.}

\item{iprint}{An integer value (0 or 1). If \code{1}, the plot will be saved to the disk. Default is \code{NULL}.}

\item{figdir}{A character string specifying the directory to save the figure. Default is \code{NULL}, which saves to the current directory if \code{iprint = 1}.}

\item{figname}{A character string specifying the name of the figure file. Default is \code{"opt_progress"}.}
}
\value{
None. The function generates and optionally saves a plot.
}
\description{
The `ccpOutVisualMJ` function visualizes the projected angles for each data block
over optimization iterations. The function allows for saving the plot as a PNG file
to the specified directory.
}
