% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIVAS-package.R
\docType{package}
\name{DIVAS-package}
\alias{DIVAS}
\alias{DIVAS-package}
\title{DIVAS: Data Integration via Analysis of Subspaces}
\description{
Based on 'Data integration via analysis of subspaces (DIVAS)' by Prothero et al. (2024), this is a package that implements the DIVAS methodology for multi-modal data integration. This package provides tools for signal extraction, noise estimation and joint structure identification from high-dimensional datasets, useful for multi-omics and other complex data types.
}
\author{
\strong{Maintainer}: Jiadong Mao \email{chiatungmao@gmail.com} (\href{https://orcid.org/0000-0002-3818-1981}{ORCID})

Authors:
\itemize{
  \item Yinuo Sun \email{yinuo.sun.1@unimelb.edu.au}
}

}
\keyword{internal}
