% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIVASmain.R
\name{DIVASmain}
\alias{DIVASmain}
\title{Data integration via analysis of subspaces}
\usage{
DIVASmain(
  datablock,
  nsim = 400,
  iprint = TRUE,
  colCent = FALSE,
  rowCent = FALSE,
  figdir = NULL
)
}
\arguments{
\item{datablock}{A list of matrices with the same number of columns (samples).}

\item{nsim}{Number of bootstrap resamples for inferring angle bounds.}

\item{iprint}{Whether to print diagnostic figures.}

\item{colCent}{Whether to column centre the input data blocks.}

\item{rowCent}{Whether to row centre the input data blocks.}

\item{figdir}{If not NULL, then diagnostic plots will be saved to this directory.}
}
\value{
A list containing DIVAS integration results. Most important ones include
  \describe{
    \item{matBlocks}{List of scores representing shared and partially shared joint structures.}
    \item{matLoadings}{List of loadings linking features in each data block with scores.}
    \item{keyIdxMap}{Mapping between indices of the previous lists and data blocks.}
  }
  See Details for more explanations.
}
\description{
Main function for DIVAS analysis. Given a list of data blocks with matched columns (samples), will return identified joint
structure with diagnostic plots.
}
\details{
DIVASmain returns a list containing all important information returned from the DIVAS algorithm.
For users, the most important ones are scores (matBlocks), loadings (matLoadings) and an index
book (keyIdxMap) explaining what joint structures each score or loading matrix correpsond to.

matBlocks is a list containing scores. Each element of matBlocks is indexed by a number.
For example, suppose one of the indices is "7", then keyIdxMap[["7"]] contains indices of data blocks
corresponding to the index 7. That is, matBlocks[["7"]] contains the scores for all samples
representing the joint structures of data blocks in keyIdxMap[["7"]].
}
\references{
Prothero, J., Jiang, M., Hannig, J., Tran-Dinh, Q., Ackerman, A. and Marron, J. S. (2024).
Data integration via analysis of subspaces (DIVAS). Test.
}
