% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJIVEAngleDiagnosticJP.R
\name{DJIVEAngleDiagnosticJP}
\alias{DJIVEAngleDiagnosticJP}
\title{Create Diagnostic Plots for DJIVE Analysis}
\usage{
DJIVEAngleDiagnosticJP(datablock, dataname, outstruct, randseed, titlestr)
}
\arguments{
\item{datablock}{List of data blocks}

\item{dataname}{Vector of data block names}

\item{outstruct}{Output structure from DJIVE analysis}

\item{randseed}{Random seed for reproducibility}

\item{titlestr}{Title string for plots}
}
\value{
List of three ggplot objects: rank, score, and loading plots
}
\description{
Create Diagnostic Plots for DJIVE Analysis
}
