% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJIVEJointStrucEstimateJP.R
\name{DJIVEJointStrucEstimateJP}
\alias{DJIVEJointStrucEstimateJP}
\title{Estimate full and partially shared joint structures
Establish a DC programming problem to estimate each partially joint structure using Penalty CCP algorithm.}
\usage{
DJIVEJointStrucEstimateJP(
  VBars,
  UBars,
  phiBars,
  psiBars,
  rBars,
  dataname = NULL,
  theta0 = 45,
  optArgin = list(),
  iprint = F,
  figdir = NULL
)
}
\arguments{
\item{VBars}{List of adjusted signal row spaces for each data block.}

\item{UBars}{List of adjusted signal column spaces for each data block.}

\item{phiBars}{Vector of adjusted perturbation angles for each data matrix.}

\item{psiBars}{Vector of loadings perturbation angles for each data matrix.}

\item{rBars}{Vector of adjusted signal ranks for each data block.}

\item{dataname}{Names of data blocks.}

\item{theta0}{Initial value for angle.}

\item{optArgin}{Aditional tuning parameters for optimisation.}

\item{iprint}{Print the figures or not.}

\item{figdir}{If not `NULL`, will be parsed as directory for storing the figures.}
}
\value{
A list containing:
  \describe{
    \item{outMap}{Mapping between joint block index set and estimated partially shared structure.}
    \item{keyIdxMap}{Mapping between joint block index and data blocks.}
    \item{anglesMap}{Mapping between joint block index and angle estimates.}
    \item{jointBlockOrder}{Record of the order of joint blocks.}
  }
}
\description{
Estimate full and partially shared joint structures
Establish a DC programming problem to estimate each partially joint structure using Penalty CCP algorithm.
}
