% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJIVEJointStrucEstimateJPLoadInfo.R
\name{DJIVEJointStrucEstimateJPLoadInfo}
\alias{DJIVEJointStrucEstimateJPLoadInfo}
\title{Estimate Joint and Partially Joint Structure}
\usage{
DJIVEJointStrucEstimateJPLoadInfo(
  VBars,
  UBars,
  phiBars,
  psiBars,
  rBars,
  datablock,
  iprint = FALSE,
  figdir = ""
)
}
\arguments{
\item{VBars}{List of adjusted signal row spaces.}

\item{UBars}{List of adjusted signal column spaces for each data block.}

\item{phiBars}{Numeric vector of perturbation angles for each data matrix.}

\item{psiBars}{Numeric vector of loadings perturbation angles for each data matrix.}

\item{rBars}{A numeric vector of adjusted signal ranks for each data matrix.}

\item{datablock}{List of data blocks.}

\item{iprint}{Logical. Whether to print results.}

\item{figdir}{Directory for saving figures. Set to `NULL` if do not want to save.}
}
\value{
A list containing:
\itemize{
  \item \code{outMap}: A mapping between joint block index sets and estimated partially shared structures.
  \item \code{keyIdxMap}: A mapping between joint block indices and data blocks.
  \item \code{jointBlockOrder}: A character vector recording the order of the joint blocks.
}
}
\description{
This function establishes a Difference of Convex (DC) programming problem to estimate each partially
joint structure. It uses the Penalty Convex-Concave Procedure (CCP) algorithm to solve the DC programming problem.
}
\details{
The function solves the DC programming problem iteratively, leveraging the Penalty CCP algorithm
to optimize the estimation of joint and partially joint structures. This approach is particularly
useful for complex multi-block data structures.
}
