% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJIVEReconstructMJ.R
\name{DJIVEReconstructMJ}
\alias{DJIVEReconstructMJ}
\title{DJIVEReconstructMJ - Reconstruct joint blocks from data blocks}
\usage{
DJIVEReconstructMJ(
  datablock,
  dataname,
  outMap,
  keyIdxMap,
  jointBlockOrder,
  doubleCenter
)
}
\arguments{
\item{datablock}{List of d_k x n data matrices.}

\item{dataname}{Vector of data matrix names.}

\item{outMap}{List mapping between block index set and estimated partially shared structure.}

\item{keyIdxMap}{List mapping between joint block index and data blocks.}

\item{jointBlockOrder}{Vector to record the joint block order.}

\item{doubleCenter}{Flag indicating whether to perform double centering.}
}
\value{
A list containing the joint basis map, loadings, and joint blocks.
}
\description{
This function reconstructs joint blocks from the provided data matrices
using the estimated partially shared structures.
}
