% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointStrucEstUtils.R
\name{Idx2numMJ}
\alias{Idx2numMJ}
\title{Convert Logical Index Array to Number (Binary to Decimal)}
\usage{
Idx2numMJ(blockIn)
}
\arguments{
\item{blockIn}{A logical vector indicating which block index is in (TRUE/FALSE).}
}
\value{
A numeric value representing the decimal conversion of the binary input.
}
\description{
This function takes a logical vector and converts it to a decimal number,
interpreting the logical values as a binary representation.
}
