% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_shrinkage.R
\name{incMarPas}
\alias{incMarPas}
\title{Incomplete Marčenko-Pastur Distribution Function}
\usage{
incMarPas(x0, beta, gamma)
}
\arguments{
\item{x0}{A numeric value representing the lower limit of integration.}

\item{beta}{A numeric value indicating the ratio of columns to rows in the data matrix.}

\item{gamma}{A numeric value specifying the power to which the function should be raised during integration.}
}
\value{
A numeric value of the integrated Marčenko-Pastur function.
}
\description{
Computes the incomplete Marčenko-Pastur distribution function, which is used in estimating noise levels.
}
