% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatReconstructMJ.R
\name{MatReconstructMJ}
\alias{MatReconstructMJ}
\title{MatReconstructMJ - Reconstruct joint block matrices and their loadings from data}
\usage{
MatReconstructMJ(X, matJointV, matJointOrder, matJointRanks)
}
\arguments{
\item{X}{d x n data matrix}

\item{matJointV}{n x sum(r_t) joint structure basis matrix}

\item{matJointOrder}{List that keeps the order of joint blocks in matJointV}

\item{matJointRanks}{Vector containing the rank of each joint block in matJointV}
}
\value{
A list containing `matBlockMap` (joint block matrix map) and `matLoadingMap` (joint block loading map)
}
\description{
MatReconstructMJ - Reconstruct joint block matrices and their loadings from data
}
