% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_shrinkage.R
\name{MedianMarcenkoPastur}
\alias{MedianMarcenkoPastur}
\title{Calculate the Median of the Marčenko-Pastur Distribution}
\usage{
MedianMarcenkoPastur(beta)
}
\arguments{
\item{beta}{A numeric value between 0 and 1 representing the aspect ratio of the matrix (\code{n/d}),
where \code{n} is the number of columns and \code{d} is the number of rows.
This ratio defines the shape of the Marčenko-Pastur distribution.}
}
\value{
A numeric value representing the estimated median of the Marčenko-Pastur distribution.
}
\description{
This function estimates the median of the Marčenko-Pastur distribution given a ratio parameter \code{beta}.
The Marčenko-Pastur distribution describes the asymptotic behavior of eigenvalues of large random matrices
and is widely used in random matrix theory and statistics for noise estimation and signal processing.
}
\details{
The function iteratively narrows down the interval containing the median of the Marčenko-Pastur
distribution by evaluating the distribution function until the interval is sufficiently small.
The bounds of the distribution are defined by \code{(1 - sqrt(beta))^2} and \code{(1 + sqrt(beta))^2}.
}
