% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalExtractOptimalShrinkage.R
\name{optimal_shrinkage}
\alias{optimal_shrinkage}
\title{Optimal Shrinkage of Singular Values}
\usage{
optimal_shrinkage(singvals, beta, loss, percentile, sigma)
}
\arguments{
\item{singvals}{A numeric vector of singular values to be shrunk.}

\item{beta}{A numeric value representing the ratio of dimensions (min(n, d) / max(n, d)), where n and d are the dimensions of the matrix.}

\item{loss}{A string specifying the loss function to be used for shrinkage. Options are "fro" (Frobenius norm), "op" (operator norm), and "nuc" (nuclear norm).}

\item{percentile}{A numeric value representing the percentile used for noise level estimation if `sigma` is not provided.}

\item{sigma}{Optional. A numeric value representing the noise level. If not provided, it will be estimated using the Median Marcenko-Pastur distribution.}
}
\value{
A list containing:
\describe{
  \item{singvals}{The shrunk singular values as a numeric vector.}
  \item{noiselvl}{The estimated noise level used in the shrinkage process.}
}
}
\description{
The `optimal_shrinkage` function performs optimal shrinkage of singular values based on the specified loss function and estimated noise level. The function supports three types of loss functions: Frobenius ("fro"), operator norm ("op"), and nuclear norm ("nuc").
}
