% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointStrucEstUtils.R
\name{projAngleMJ}
\alias{projAngleMJ}
\title{Calculate the Projection Angle between a Vector and a Basis Matrix}
\usage{
projAngleMJ(x, V)
}
\arguments{
\item{x}{A numeric vector of length \code{n}, representing an \eqn{n \times 1} vector.}

\item{V}{A numeric matrix of dimension \eqn{n \times r}, representing a basis matrix of the subspace.}
}
\value{
A numeric value representing the projected angle of \code{x} on \code{V}, in degrees.
The value is in the range \eqn{[0, 180]}.
}
\description{
The function `projAngleMJ` calculates the angle between a given vector and its projection
onto the subspace spanned by a basis matrix. The resulting angle provides a measure of
how closely the vector aligns with the subspace.
}
