% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJIVEAngleDiagnosticJP.R,
%   R/SignalExtractOptimalShrinkage.R
\name{randDirAngleMJ}
\alias{randDirAngleMJ}
\title{Calculate Random Direction Angles}
\usage{
randDirAngleMJ(n, r, nsim)

randDirAngleMJ(n, r, nsim)
}
\arguments{
\item{n}{An integer representing the dimension of the vector space.}

\item{r}{An integer representing the dimension of the subspace.}

\item{nsim}{An integer representing the number of simulation samples.}
}
\value{
Vector of angles in degrees

A numeric vector of length `nsim`, containing the simulated random direction angles (in degrees).
}
\description{
Computes angles between random vectors and their projections onto first r components

This function simulates the angles between a random direction and a fixed rank `r` subspace in \eqn{R^n}.
Specifically, it calculates the angles between a random vector and the subspace spanned by the first `r`
dimensions in an \eqn{n}-dimensional space.
}
