% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJIVESignalExtractJP.R
\name{DJIVESignalExtractJP}
\alias{DJIVESignalExtractJP}
\title{Signal Matrix Extraction for DIVAS}
\usage{
DJIVESignalExtractJP(
  datablock,
  nsim = 400,
  iplot = FALSE,
  colCent = F,
  rowCent = F,
  cull = 0.382,
  noisepercentile = rep(0.5, length(datablock)),
  noiselvls = NULL
)
}
\arguments{
\item{datablock}{A list of feature by sample (eg gene by cell) data matrices, each matrix corresponding to a data block. Matrices have to have same number of columns (mathced samples).}

\item{nsim}{An integer specifying the number of bootstrap samples.}

\item{iplot}{A logical to indicate whether plots should be generated for visualizing singular value shrinkage.}

\item{colCent}{A logical indicating whether columns should be centered.}

\item{rowCent}{A logical indicating whether rows should be centered.}

\item{cull}{A numeric value for the culling parameter to adjust signal rank.}

\item{noisepercentile}{A numeric vector specifying the percentiles used for noise estimation for each data block.}

\item{noiselvls}{A list specifying noise levels for each data block; if NULL, noise levels are estimated internally.}
}
\value{
A list containing:
  \describe{
    \item{VBars}{List of adjusted signal row spaces for each data block.}
    \item{UBars}{List of adjusted signal column spaces for each data block.}
    \item{phiBars}{Vector of adjusted perturbation angles for each data matrix.}
    \item{psiBars}{Vector of loadings perturbation angles for each data matrix.}
    \item{rBars}{Vector of adjusted signal ranks for each data block.}
    \item{EHats}{List of estimated noise matrices for each data block.}
    \item{singVals}{List of singular values before shrinkage for each data block.}
    \item{singValsHat}{List of singular values after shrinkage for each data block.}
    \item{rSteps}{List of signal rank adjustment steps for each data block.}
    \item{VVHatCacheBars}{List of cached VVHat matrices for bootstrap samples.}
    \item{UUHatCacheBars}{List of cached UUHat matrices for bootstrap samples.}
  }
}
\description{
Extracts the signal matrix for each data block, estimates signal ranks, and adjusts perturbation angles for each data matrix.
}
