% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_shrinkage.R
\name{randDirAngleMJ}
\alias{randDirAngleMJ}
\title{Simulate Random Direction Angles in Vector Space}
\usage{
randDirAngleMJ(n, r, nsim)
}
\arguments{
\item{n}{An integer representing the dimension of the vector space.}

\item{r}{An integer representing the dimension of the subspace.}

\item{nsim}{An integer representing the number of simulation samples.}
}
\value{
A numeric vector of length `nsim`, containing the simulated random direction angles (in degrees).
}
\description{
This function simulates the angles between a random direction and a fixed rank `r` subspace in \eqn{R^n}.
Specifically, it calculates the angles between a random vector and the subspace spanned by the first `r`
dimensions in an \eqn{n}-dimensional space.
}
